+++ frontend/src/utils/positionUtil.ts
